+++ custom_components/youtube_pro/const.py
@@ -22,6 +22,8 @@ SERVICE_SET_TIMER = "set_timer"
 SERVICE_START_RADIO = "start_radio"
 SERVICE_PLAY_PERSONAL_MIX = "play_personal_mix"
 SERVICE_LISTENER_FEEDBACK = "listener_feedback"
+SERVICE_PAUSE = "pause"
+SERVICE_RESUME = "resume"
 
 REPEAT_MODES = ("off", "all", "one")
 TIMER_TYPES = ("play", "stop")